refactor: push 패키지에서 teambottari 관련 파일들을 의존하지 않도록 패키지 이동
