--- backend/bottari/src/main/java/com/bottari/push/fcm/dto/MessageType.java
@@ -1,10 +1,0 @@
-package com.bottari.push.fcm.dto;
-
-public enum MessageType {
-
-    TEAM_ITEM_CHANGED,
-    REMIND_BY_ITEM,
-    REMIND_BY_TEAM_MEMBER,
-    EXIT_TEAM_BOTTARI;
-
-}
