--- backend/bottari/src/main/java/com/bottari/push/fcm/dto/SendMessageRequest.java
@@ -1,13 +1,0 @@
-package com.bottari.push.fcm.dto;
-
-import java.util.Map;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class SendMessageRequest {
-
-    private final Map<String, String> data;
-    private final MessageType messageType;
-}
