+++ backend/bottari/src/main/java/com/bottari/push/sse/SseChannel.java
@@ -1,4 +1,4 @@
-package com.bottari.push.sse.component;
+package com.bottari.push.sse;
 
 import com.bottari.push.ChannelType;
 import com.bottari.push.ConnectionBasedChannel;
