+++ backend/bottari/src/main/java/com/bottari/push/sse/SseRepository.java
@@ -1,4 +1,4 @@
-package com.bottari.push.sse.component;
+package com.bottari.push.sse;
 
 import java.util.List;
 import java.util.Map;
