+++ backend/bottari/src/main/java/com/bottari/teambottari/adapter/data/CreateAssignedItemData.java
@@ -1,4 +1,4 @@
-package com.bottari.push.sse.dto;
+package com.bottari.teambottari.adapter.data;
 
 import com.bottari.teambottari.event.CreateAssignedItemEvent;
 import java.time.LocalDateTime;
