+++ backend/bottari/src/main/java/com/bottari/teambottari/adapter/data/DeleteTeamSharedItemData.java
@@ -1,4 +1,4 @@
-package com.bottari.push.sse.dto;
+package com.bottari.teambottari.adapter.data;
 
 import com.bottari.teambottari.event.DeleteTeamSharedItemEvent;
 import java.time.LocalDateTime;
