+++ backend/bottari/src/test/java/com/bottari/teambottari/service/TeamBottariServiceTest.java
@@ -9,7 +9,7 @@ import com.bottari.fixture.MemberFixture;
 import com.bottari.fixture.TeamBottariFixture;
 import com.bottari.member.domain.Member;
 import com.bottari.push.PushManager;
-import com.bottari.push.fcm.TeamBottariMessageConverter;
+import com.bottari.teambottari.adapter.TeamBottariMessageConverter;
 import com.bottari.teambottari.domain.TeamAssignedItem;
 import com.bottari.teambottari.domain.TeamAssignedItemInfo;
 import com.bottari.teambottari.domain.TeamBottari;
